Code was refactored;
Now import & export tasks are deligated to a class
implementing TableExportUtility interface; this way
MetadataTable can be in 2 states, depending on the format
of the table which is used to import/export the data (either long-table or wide-table);

The next step is to implement the WideTableExportUtility class;
