+++ src/main/java/io/github/mzmine/datamodel/MZmineProject.java
@@ -23,7 +23,7 @@ import io.github.mzmine.modules.io.projectload.CachedIMSRawDataFile;
 import io.github.mzmine.parameters.UserParameter;
 import io.github.mzmine.project.impl.ProjectChangeEvent.Type;
 import io.github.mzmine.project.impl.ProjectChangeListener;
-import io.github.mzmine.project.parameterssetup.MetadataTable;
+import io.github.mzmine.project.parameterssetup.table.MetadataTable;
 import io.github.mzmine.util.spectraldb.entry.SpectralLibrary;
 import java.io.File;
 import java.util.Hashtable;
