+++ src/main/java/io/github/mzmine/project/parameterssetup/ProjectParametersSetupDialogController.java
@@ -26,10 +26,11 @@ import io.github.mzmine.main.MZmineCore;
 import io.github.mzmine.parameters.parametertypes.ComboParameter;
 import io.github.mzmine.parameters.parametertypes.StringParameter;
 import io.github.mzmine.project.parameterssetup.ProjectMetadataParameters.AvailableTypes;
-import io.github.mzmine.project.parameterssetup.columns.DoubleMetadataColumn;
-import io.github.mzmine.project.parameterssetup.columns.DateMetadataColumn;
-import io.github.mzmine.project.parameterssetup.columns.MetadataColumn;
-import io.github.mzmine.project.parameterssetup.columns.StringMetadataColumn;
+import io.github.mzmine.project.parameterssetup.table.columns.DoubleMetadataColumn;
+import io.github.mzmine.project.parameterssetup.table.columns.DateMetadataColumn;
+import io.github.mzmine.project.parameterssetup.table.columns.MetadataColumn;
+import io.github.mzmine.project.parameterssetup.table.columns.StringMetadataColumn;
+import io.github.mzmine.project.parameterssetup.table.MetadataTable;
 import io.github.mzmine.util.ExitCode;
 import java.io.File;
 import java.time.LocalDateTime;
