+++ src/main/java/io/github/mzmine/project/parameterssetup/table/columns/DateMetadataColumn.java
@@ -16,35 +16,44 @@
  *
  */
 
-package io.github.mzmine.project.parameterssetup.columns;
+package io.github.mzmine.project.parameterssetup.table.columns;
 
 import io.github.mzmine.project.parameterssetup.ProjectMetadataParameters.AvailableTypes;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeParseException;
 
 /**
- * Specific String-type implementation of the project parameter.
+ * Specific Date-type implementation of the project parameter.
  */
-public final class StringMetadataColumn extends MetadataColumn<String> {
+public final class DateMetadataColumn extends MetadataColumn<LocalDateTime> {
 
-  public StringMetadataColumn(String title) {
+  public DateMetadataColumn(String title) {
     super(title);
   }
 
-  public StringMetadataColumn(String title, String description) {
+  public DateMetadataColumn(String title, String description) {
     super(title, description);
   }
 
   @Override
   public boolean checkInput(Object value) {
-    return value instanceof String;
+    return value instanceof LocalDateTime;
   }
 
   @Override
   public AvailableTypes getType() {
-    return AvailableTypes.TEXT;
+    return AvailableTypes.DATETIME;
   }
 
   @Override
-  public String convert(String input, String defaultValue) {
-    return input == null ? defaultValue : input.trim();
+  public LocalDateTime convert(String input, LocalDateTime defaultValue) {
+    try {
+      //ISO-8601 format
+      return input == null ? defaultValue : LocalDateTime.parse(input.trim());
+//      return LocalDateTime.parse(input.trim(), DateTimeFormatter.ofPattern("dd-MM-yyyy HH:mm:ss"));
+    } catch (DateTimeParseException ignored) {
+    }
+
+    return null;
   }
 }
